refactor: rename and replace hyperLiquidDappDetect with hyperLiquidDappDetecter
- Replaced all instances of hyperLiquidDappDetect with hyperLiquidDappDetecter for consistency and clarity.
- Deleted the old hyperLiquidDappDetect file and moved its functionality to the new hyperLiquidDappDetecter file.
- Updated related imports and method calls to ensure proper functionality.

This update enhances code maintainability and improves naming conventions within the built-in Perp injection system.
