+++ packages/providers/inpage-providers-hub/src/builtInPerpInjected/hyperLiquidDappDetecter.ts
